% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{add_enzymeInfo}
\alias{add_enzymeInfo}
\title{Add the information of enzymes to the dataframe of reactions}
\usage{
add_enzymeInfo(tmpr, rbpo, enzymeClass, verbose = TRUE)
}
\arguments{
\item{tmpr}{A dataframe created by function "extract_reactionInfo", containing the backbone of reactions}

\item{rbpo}{A rBioPax object generated by reading an owl file using package "rBiopaxParser".}

\item{enzymeClass}{A character string, describing the type of desired class. Available classes can be searched in rbpo$dt}

\item{verbose}{Logical, TRUE or FALSE. Whether to print  messages.}
}
\value{
A dataframe with enzyme info of desired type of enzymes.
}
\description{
This function extracts the desired type of enzyme ids from the rBioPax object, and add them to the reactions
}
\details{
This function searches for the ids of a given enzyme type, then track down the reaction id catalyzed by the enzyme, then add the info of the enzyme to the reaction. The enzyme types are one of "Catalysis" or "Control".
}
