% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{add_modulationInfo}
\alias{add_modulationInfo}
\title{Add the information of modulator to the dataframe of reactions}
\usage{
add_modulationInfo(tmpr, rbpo, enzymeClass = "Modulation", verbose = TRUE)
}
\arguments{
\item{tmpr}{A dataframe created by function "extract_reactionInfo", containing the information of reactions}

\item{rbpo}{A rBioPax object generated by reading an owl file using package "rBiopaxParser".}

\item{enzymeClass}{A character string, describing the type of desired class. Default is "Modulation". Will expand additional functions in future.}

\item{verbose}{Logical, TRUE or FALSE. Whether to print  messages.}
}
\value{
A dataframe with modulation's info for reactions.
}
\description{
This function extracts the modulators from the rBioPax object if exist, and add them to the reactions
}
\details{
This function searches for the ids of modulator, then track down the reaction id catalyzed by the enzyme, and add the info of the enzyme to the reaction.

Modulation is a distinct class compared to "Catalysis" and "Control", since it is defined to record enzymes that change the state of another enzyme which catalyze a reaction.

Some databases use "Modulation", while some use "Control" to store the same reactions.
}
