% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{build_component_structure}
\alias{build_component_structure}
\title{Build a vector containing the components of ids.}
\usage{
build_component_structure(myvec, rbpo)
}
\arguments{
\item{myvec}{A vector containing the ids which need to be decomposed}

\item{rbpo}{A rBioPax object generated by reading an owl file using package "rBiopaxParser".}
}
\value{
A vector containing the decomposed ids or original ids if they are not able to be decomposed.
}
\description{
This function break down the ids in a given vector into their components recursively.
}
\details{
This function takes a vector of ids. For each id, this function tries to break it down into its components. Sometimes an id is not a complex, but a "set". A set will also be broken down into its elements.
}
