% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{build_cplx2cpnt}
\alias{build_cplx2cpnt}
\title{A wrapper building the two-column matching table between complexes and components/elements.}
\usage{
build_cplx2cpnt(hbiopax, out.cplx2cpnt)
}
\arguments{
\item{hbiopax}{A rBioPax object created by function "read_owl".}

\item{out.cplx2cpnt}{A character string specifying the path and name of the output RData file which stores the matching table.}
}
\value{
A two-column dataframe containing the matching information between complexes and components/elements.
}
\description{
This function looks for the ids of class "Complex", and calls function "create_complexComponentTable" to build a complex to components matching table.
}
\details{
This function checks the existence of the output RData file, if it exists, then load the R object from the file directly, if not, then go on to build the matching table.

Sometimes there's no complexes in the rBioPax object, then this function will return a void dataframe.
}
