% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{build_rawSIF}
\alias{build_rawSIF}
\title{A wrapper building the GIN in SIFI format with local unreplaced ids.}
\usage{
build_rawSIF(out.sif, reactions.full, cplx2cpnt)
}
\arguments{
\item{out.sif}{A character string specifying the path and name of the output RData file containing the GIN with local ids.}

\item{reactions.full}{A dataframe created by function "build_reactionsFull" which contains all the information for all reactions.}

\item{cplx2cpnt}{A two-column dataframe containing the matching information between complexes and their components/elements.}
}
\value{
A three-column dataframe in SIFI format, with starting nodes, ending nodes and edge properties.
}
\description{
This function build a raw dataframe in SIFI format based on the reactions dataframe and the complex matching table.
}
\details{
This a wrapper building the GIN in SIFI format with local unreplaced ids. It will first check the existence of already-built raw SIFI file. If not exsiting, then it will call the function "convert_toSIF", then save the results into a local file and return the dataframe.
}
