% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{build_reactionsFull}
\alias{build_reactionsFull}
\title{A wrapper building the dataframe which stores all of the information for all reactions from a rBioPax object.}
\usage{
build_reactionsFull(out.reactions.full, hbiopax, verbose = TRUE)
}
\arguments{
\item{out.reactions.full}{A character string specifying the name and location of an existing .RData file.}

\item{hbiopax}{A rBioPax object created by function "read_owl".}

\item{verbose}{Logical for whether or not to report warnings.}
}
\value{
A dataframe containing all of the reactions from a rBioPax object, with enzymes' and modulators' information.
}
\description{
This function extracts all of the desired reactions from a rBioPax object into a dataframe, then fill the dataframe with the information of enzymes and modulators.
}
\details{
This function will first build a dataframe by extracting the reactions from classes "BiochemicalReaction", "Degradation", "Conversion", "ComplexAssembly", "Transport", "TransportWithBiochemicalReaction", and "TemplateReaction".

Then it will extract the enzymes from classes "Catalysis" and "Control" for each reaction. Finally, it will extract "Modulation" if there is any. Before it returns the resulting dataframe, it will also save the dataframe to a local file to save time for future loading.
}
