% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{convert_toSif}
\alias{convert_toSif}
\title{Convert the reaction table into SIFI format}
\usage{
convert_toSif(df.reactions, matchTable = cplx2cpnt)
}
\arguments{
\item{df.reactions}{A dataframe containing the information of all of the reactions of a database.}

\item{matchTable}{A dataframe which stores the matching information between complexes and components/elements}
}
\value{
A three-column dataframe containing the edges from substrates and enzymes to intermediate, and from intermediate to products.
}
\description{
This function takes the dataframe of the reactions and convert the reactions into topological network in simple interactions format with intermediates.
}
\details{
This function calls several other functions to decompose the complexes, generate intermediates, and create edges in SIFI format. The SIFI format is similar to SIF format, but with additional intermediate nodes.
}
