% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{create_complexComponentTable}
\alias{create_complexComponentTable}
\title{Build a table containing the matching information of complexes and their components}
\usage{
create_complexComponentTable(rbpo, class.this = "Complex")
}
\arguments{
\item{rbpo}{A rBioPax object generated by reading an owl file using package "rBiopaxParser".}

\item{class.this}{A character string specifying the class of the ids. Default if "Complex".}
}
\value{
A two-column dataframe containing the original ids of "Complex" class and the decomposed ids collapsed by "_".
}
\description{
This function create a dataframe by decomposing all the ids of "Complex" into components and set elements.
}
\details{
This function searches for all of the ids of "Complex", then loop for each id and decompose into components and elements, and collapse the resulting ids by "_", and finally store the matching info in a dataframe.
}
