% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{create_edgesFromReaction}
\alias{create_edgesFromReaction}
\title{Generate edges based on the information of a reaction}
\usage{
create_edgesFromReaction(
  vec.substrate,
  vec.product,
  enzyme,
  controlType,
  tmp.itmd,
  reactionID
)
}
\arguments{
\item{vec.substrate}{A vector containing the ids of substrates of a reaction.}

\item{vec.product}{A vector containing the ids of products of a reaction.}

\item{enzyme}{A character string of the enzyme id of a reaction. If there's no enzyme info for this reaction, then this value will be "NoEnzyem" or "".}

\item{controlType}{A character string specifying the control type of the enzyme to the reaction. This info can be retrieved by the function "add_enzymeInfo" when building the table of reactions.}

\item{tmp.itmd}{A character string specifying the name of the intermediate node for this reaction. Generated by function "generate_intermediate".}

\item{reactionID}{A character string of a reaction id. Should match the other parameters.}
}
\value{
A three-column dataframe containing the edges from substrates and enzymes to intermediate, and from intermediate to products.
}
\description{
This function generates edges based on the substrates, products, enzymes and intermediate nodes of a reaction.
}
\details{
This function generates edges from substrates and enzymes to intermediate, and from intermediate to products.
}
