% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{create_singleReactionBackbone}
\alias{create_singleReactionBackbone}
\title{Create a dataframe for a reaction, containing only basic information}
\usage{
create_singleReactionBackbone(rbpo, id_single, bpclass)
}
\arguments{
\item{rbpo}{A rBioPax object generated by reading an owl file using package "rBiopaxParser".}

\item{id_single}{A character string, which is a single reaction id in the specific format defined by the owl file.}

\item{bpclass}{A character string, which is an annotation of the type of the reaction}
}
\value{
A small dataframe containing the information of a single reaction. The information includes "left", "right", "direction", "name", "enzyme", "controlType", "modulator", "modulatorType".
}
\description{
This function create a dataframe given a reaction id from the rbiopax object
}
\details{
The rbpo$dt is fundamentally a dataframe. The information of a reaction is retrieved by subsetting the dataframe restricting the "id" slot to be "id_single".

The reaction will have "left" and "right" as the substrates and products; "direction"; "name" as the displayName

Other slots: "enzyme", "controlType","modulator",and "modulatorType" are only initialized with default value.
}
