% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{decomplex}
\alias{decomplex}
\title{Decompose a complex id into its components}
\usage{
decomplex(cplx, rbpo)
}
\arguments{
\item{cplx}{An internal id from "rbpo"}

\item{rbpo}{A rBioPax object generated by reading an owl file using package "rBiopaxParser".}
}
\value{
A vector containing the decomposed ids or original ids if they are not able to be decomposed.
}
\description{
This function break down an id, calling the function "build_component_structure"
}
\details{
This function takes an id, searches for components and set elements, then calling "build_component_structure" to decompose recursively.
}
