% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{extract_reactionInfo}
\alias{extract_reactionInfo}
\title{Retrieve all the intended class of reactions from a rBioPax object}
\usage{
extract_reactionInfo(rbpo, bpclass)
}
\arguments{
\item{rbpo}{A rBioPax object generated by reading an owl file using package "rBiopaxParser".}

\item{bpclass}{A character string, describing the type of desired class. Available classes can be searched in rbpo$dt}
}
\value{
This function returns a dataframe, containing the substrate and products of each reactions.
}
\description{
This function takes a rBioPax object and the type of desired class, then build a backbone of dataframe containing all the reactions without enzyme information.
}
\details{
This is a wrapper to loop for all the reaction ids, creating a dataframe of basic information of reactions
}
