% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{generate_intermediate}
\alias{generate_intermediate}
\title{Generate intermediate nodes for a reaction.}
\usage{
generate_intermediate(
  vec.substrate,
  enzyme,
  controlType,
  modulator,
  modulateType,
  complexTable,
  reactionID
)
}
\arguments{
\item{vec.substrate}{A vector containing the ids of substrates of a reaction.}

\item{enzyme}{A character string of the enzyme id of a reaction. If there's no enzyme info for this reaction, then this value will be "NoEnzyem" or "".}

\item{controlType}{A character string specifying the control type of the enzyme to the reaction. This info can be retrieved by the function "add_enzymeInfo" when building the table of reactions.}

\item{modulator}{A character string of the modulator id. A modulator changes the state of an enzyme which catalyze a reaction.}

\item{modulateType}{A character string specifying the control type of the modulator to the enzyme. Generally it's either a positive or a negative relation.}

\item{complexTable}{A two-column dataframe which provides matching info between complexes and their components or elements.}

\item{reactionID}{A character string of a reaction id. Should match the other parameters except "complexTable".}
}
\value{
A list containing the intermediate node of the reaction. If there is a modulator in the reaction, then the list will have a second element which stores the edges describing the interactions between the modulator, enzyme and the reaction.
}
\description{
This function generates intermediate nodes.
}
\details{
This function aggregates the information of substrates and enzyme of a reaction and forms the name of an intermediate node. If a modulator presents in the reaction, then this function will create additional intermedate nodes for the reactions between enzyme and modulator, and store the edges in the second element of the result list. Note the intermediate node of enzyme;modulator will point to the reaction's intermediate node (substrates;enzyme) to avoid ambiguity.
}
