% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{read_owl}
\alias{read_owl}
\title{A wrapper for reading the owl file calling the function "readBiopax" from rBioPaxParser}
\usage{
read_owl(owl, out.biopax, flag)
}
\arguments{
\item{owl}{A character string specifying the name and location of an owl file.}

\item{out.biopax}{A character string specifying the name and location of a RData file of rBioPax object.}

\item{flag}{A character string specifying the state of an variable "input". If input has been initialized, then input should be a string of the name of the owl file (without path). If not, then input will be set to "noFile" before passed to this function.}
}
\value{
A rBioPax object loaded from file "owl", and will create a .RData file for this object in local dir.
}
\description{
This function checks whether there is already a RData file of rBioPax object, if yes, load the existing file, if no, load the given owl file.
}
\details{
If the input owl file name has not been specified in the command line, then this function will cancel the whole process if there's no existing .RData file of the rBioPax object. So it's safer to specify the "--input" option when calling from command line. After loading the owl file, this function will save the object to a file named as "out.biopax" to save time for future loading.
}
