% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{replace_complexWithComponent}
\alias{replace_complexWithComponent}
\title{Replace the ids in a vector with the decomposed id strings}
\usage{
replace_complexWithComponent(vec.in, matchTable = cplx2cpnt)
}
\arguments{
\item{vec.in}{A vector containing molecular ids.}

\item{matchTable}{A two-column dataframe containing the original complex ids and the decomposed id strings. The table can be created by "create_complexComponentTable".}
}
\value{
A vector of ids which are decomposed into components or elements.
}
\description{
This function try to search in the complex to component matching table to replace the ids in a given vector.
}
\details{
This function replaces the ids in the given vector with decomposed id strings, using the match table created by "create_complexComponentTable".
}
