% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_SIFItools.R
\name{replace_id_with_annotation}
\alias{replace_id_with_annotation}
\title{Replace internal ids with external ids}
\usage{
replace_id_with_annotation(
  biopax,
  sif,
  start,
  end,
  matchTable,
  globalReplace.before = c("_", "-"),
  globalReplace.after = c(" ", "-"),
  nreport = 1,
  verbose = TRUE
)
}
\arguments{
\item{biopax}{A rBioPax object generated by reading an owl file using package "rBiopaxParser".}

\item{sif}{A dataframe in SIFI format, containing the edges with internal ids.}

\item{start}{An integar specifying the starting line of the looping. Useful for paralleling the processing of large databases. Should be smaller than end.}

\item{end}{An integar specifying the ending line of the looping. Useful for paralleling the processing of large databases. Should be bigger than start.}

\item{matchTable}{A two-column dataframe, with the first column containing the strings that need to be replaced, and the second column containing the strings after replacement. Example: replace "uniprot:" with "UniProt".}

\item{globalReplace.before}{A vector of two elements, specifying the global treatment of special characters which are not desired. This action will be taken BEFORE the replacement of internal ids with external ids. Default (c("_","-")) is to replace "_" with "-".}

\item{globalReplace.after}{A vector of two elements, specifying the global treatment of special characters which are not desired. This action will be taken AFTER the replacement of internal ids with external ids. Default (c(" ","-")) is to replace " " with "-".}

\item{nreport}{An integar specifying the limit number of candidate annotations for reporting. Will be passed to function "get_id_annotation".}

\item{verbose}{Logical, TRUE or FALSE. Whether to report warnings.}
}
\value{
A three-column dataframe containing the edges from substrates and enzymes to intermediate, and from intermediate to products, using external ids.
}
\description{
This function replaces the ids which is only meaningful within the use of a specific database with ids in commonly used formats (for example gene symbol and CHEBI id).
}
\details{
This function calls several other functions to replace the ids which is only meaningful within the use of a specific database with ids in commonly used formats.
}
